Goodevening my goodsir
Hi.
I have updated the code. Now it saves the distance from the wall using the lazer scan. Plots the graph of the distance to the nearest object. Plot the calculated distance to the wall (from encoder ticks and initial position)
Then the data plotted is saved in csv, so it can be studied later if needed